feat: ssr support for subscriptions
